clintrinsics
